Distributed pretraining

+++ src/xlmr_siamese/util.py
@@ -4,6 +4,6 @@ def mean_encoded_seq_batch(encoded, seq, ignore_index=1):
     encoded (batch_size,seq_len,emb_dim)
     seq (batch_size,seq_len)
     '''
-    mask = seq!=ignore_index
+    mask = (seq!=ignore_index).float()
     masked = (encoded*mask.unsqueeze(-1))
     return masked.sum(1)/mask.sum(1).unsqueeze(-1)
